+++ www/src/Controller/Course/ListCoursesJson.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Werner\BookStore\Controller;
+namespace Werner\BookStore\Controller\Course;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ObjectRepository;
